Dirty hacks for java.
By no means permanent!

Looks like we need to re-structure the benchmark hierarchy. Java has a
tendency to throw a load of files all over the filesystem. I think we
really need directories per-variant/benchmark.

Also we have to devise a way for the runner to set the java CLASSPATH.