Remove CPU pinning support.
CPU pinning seemed like a good idea, but in reality it negatively
impacts VMs which use separate threads (or processes)for (e.g.)
compilation.

When the VMs create such threads these children inherit the CPU affinity
(the CPUs on which they can run) from the parent. The result is that we
create artificial CPU contention, thus slowing the VM down.

We found that the initial iterations of some JRubyTruffle benchmarks
were impacted by a >2x slowdown.