Krun's new, and more flexible, instrumentation mode.
Passes extra arguments to the iterations runner which is then expected
to write directly into the instrumentation directory.

Before Krun would scrape stderr from the pipe it opens to the iterations
runner, before dumping it into a file on behalf of the runner. That code
still exists as we have instrumentation using it, however, this should
be ported to the new API soon.