Allocate arrays of a fixed type in Python.
On VMs like PyPy, this code:

  x = [0]
  x[0] = 0.0

causes a reallocation (from list of unboxed ints to list of unboxed floats),
even though the number of elements doesn't change -- which is precisely the
opposite of what we wanted to do.

This commit moves things to use the array module with arrays of fixed types that
can't possibly reallocate.