Check throttle flags in Linux sysfs.
Changes in throttle flags now cause a re-run.

The safest way to do this without moving around too much of the fragile
exception handling code around is to have the check in the ExecutionJob
itself (since that's where all of the re-run logic currently stems
from).

I did experiment with putting the check in the Scheduler itself, so that
the check could be closer to the other checks (e.g. dmesg), but then we
end up duplicating re-run logic.

The initial throttle flags read comes after the initial dmesg read.

The second throttle flags read comes before the second dmesg read.