Implement timeouts for process executions.
This is required because sometimes you need to benchmark VMs and/or
benchmark suites which freeze up. We've seen this numerous times with
Java VMs and benchmarks. Without having a timeout to kill frozen process
executions, you can't make progress and have to restart the experiment
(possibly skipping that benchmark).

Further, by restarting the experiment you are biasing your experiment:
giving benchmarks that have the worst performance (that's what a
non-terminating benchmark is) another chance to redeem themselves.

Timed-out process executions now show as a 'T' flag in the manifest file.
Flags for finished process executions are carried through to the results
file in a new `pexec_flags` field. This field can be one of {C, E, T}
(completed, errored, timed-out).